All files merged in one single class. Kind like the NanoHTTPD.
