--- websocket/src/main/java/fi/iki/elonen/IWebSocketFactory.java
@@ -1,7 +1,0 @@
-package fi.iki.elonen;
-
-import fi.iki.elonen.NanoHTTPD.IHTTPSession;
-
-public interface IWebSocketFactory {
-    WebSocket openWebSocket(IHTTPSession handshake);
-}
